patch to make efibootmgr have non-zero exit code on errors
We have some automated tools that use efibootmgr which are having a hard
time detecting when efibootmgr fails since it exits with 0 on many
failures.  This patch catches (most) errors and exits with non-zero.  I
also added several error messages for to make it more obvious what was
wrong with the command line arguments.

Signed-off-by: Matt Domsch <Matt_Domsch@dell.com>