From: Lane Winner<lane.winner@oracle.com>
make_boot_var does not check for failed status with create_variable.
This can result in a memory leak.
Additionally the user should be notified of the problem.

We encounter this issue on one system after filling up the UEFI boot list
with dummy devices.

The patch fix the problem. It was verified on a Mensa system using RHEL 6.0

Signed-off-by: Yinghai Lu<yinghai@kernel.org>