Add a --keep option to preserve the old BootOrder entries when using -o.
So basically if you've got a boot order of:

0000,0001,0018,001A,0019

And you do "efibootmgr -o 001A --keep", you'll get:

001A,0000,0001,0018,0019

And if you subsequently do "efibootmgr -o 0002 --keep" you'll get:

0002,001A,0000,0001,0018,0019

This is to work around a buggy firmware reported in rhbz#889509 ,
which will erroneously try to re-populate BootOrder /and create new
Boot#### entries to do so/ if you supply a BootOrder without the
defaults in it.

This firmware is garbage.

Signed-off-by: Peter Jones <pjones@redhat.com>