Make default '--loader' build-time configurable.
Each distribution uses a specific 'vendor'-directory in the
'efi' directory on the EFI System Partition to store their
EFI loader.  There's little use in hardcoding the value
for just one in 'efibootmgr', which is displayed in '--help'
and used as default without '--loader'.

Simply use
  make OS_VENDOR=<vendor> EFI_LOADER=<loader>.efi
to get the desired value.  :)

Signed-off-by: Raymund Will <rw@suse.com>