Rework raw optional loader data printing to be more useful.
Previously we decided if a character was printable on a byte-by-byte
basis, and if the character was not printable, showed "." like "hexdump -C"
does.  hexdump gets away with this because the hex is there, but we
don't have that going on.

This patch changes it to decide up front if there are *any* unprintable
characters, and if so, dump hex for the whole thing instead of printing
useless nonsense.

(Incidentally this also alphabetizes the include files, because I like
that better.)

Fixes github issue #123

Signed-off-by: Peter Jones <pjones@redhat.com>