Insert, Move, and Delete within the BootOrder
Provides the functionality from Issue #126

1) When adding a new entry with -c, optionally allow an Index to be
provided with -I. If no index is provided, it defaults to 1, which
inserts the new entry first in the boot order, and provides the
same functionality as previous versions. If -I is provided, it
specifies where in the BootOrder to insert the new entry. (If -I 2
is passed, the new entry will be inserted in the second place.)
If the index specified is greater than the number of entries in the
BootOrder, the entry is placed last.

2) In addition to modifying the BootOrder by passing an entire new
BootOrder as comma-delimted string to -o, two new formats are allowed:

   -o -5 will remove entry 5 (Boot0005) from the BootOrder string
   -o +5 will add entry 5 (Boot0005) to the BootOrder string.
         By default the entry will be added at the beginning of the
         BootOrder (in first place), but this behavior can be
         overridden by specifying an index with -I as above.
         If the specified entry already exists in the BootOrder,
         then it will be moved to the specified index.